Split up saver test as it frequently times out.
